requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
paho-mqtt==2.1.0
